pyproject.toml
[tool.poetry]
name = "mediasoft"
version = "0.1.0"
description = ""
authors = ["RezuanDzibov <rezuan.dzbov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.73.0"
SQLAlchemy = "^1.4.31"
asyncpg = "^0.25.0"
uvicorn = "^0.17.4"
alembic = "^1.7.6"
python-dotenv = "^0.19.2"
flake8 = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
